+++ MainApp/Vtodo.Entities/Models/ProjectAccountsRoles.cs
@@ -1,4 +1,3 @@
-using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using Vtodo.Entities.Enums;
 
@@ -7,10 +6,10 @@ namespace Vtodo.Entities.Models
     public class ProjectAccountsRoles
     {
         [ForeignKey("Project"), Column(Order = 0)]
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         
         [ForeignKey("Account"), Column(Order = 1)]
-        public int AccountId { get; set; }
+        public long AccountId { get; set; }
         
         public Project Project { get; set; } = null!;
