+++ MainApp/Vtodo.Entities/Models/ProjectBoardFile.cs
@@ -6,10 +6,10 @@ namespace Vtodo.Entities.Models
     public class ProjectBoardFile
     {
         [ForeignKey("Project"), Column(Order = 0)]
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         
         [ForeignKey("Board"), Column(Order = 1)]
-        public int BoardId { get; set; }
+        public long BoardId { get; set; }
         
         [Required]
         public string FileName { get; set; } = string.Empty;
