+++ MainApp/Vtodo.Entities/Models/ProjectTaskFile.cs
@@ -6,10 +6,10 @@ namespace Vtodo.Entities.Models
     public class ProjectTaskFile
     {
         [ForeignKey("Project"), Column(Order = 0)]
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         
         [ForeignKey("TaskM"), Column(Order = 1)]
-        public int TaskId { get; set; }
+        public long TaskId { get; set; }
         
         [Required]
         public string FileName { get; set; } = string.Empty;
