+++ MainApp/Vtodo.Infrastructure.Implementation/Services/CurrentAccountService.cs
@@ -29,7 +29,7 @@ namespace Vtodo.Infrastructure.Implementation.Services
         {
             if (!(_httpContextAccessor.HttpContext.User?.Identity?.IsAuthenticated ?? false)) throw new UnauthorizedException();
             
-            int.TryParse(_httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value, out var userId);
+            long.TryParse(_httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value, out var userId);
 
             return _dbContext.Accounts.FirstOrDefault(x => x.Id == userId) ?? throw new UnauthorizedException();
         }
