+++ MainApp/Vtodo.UseCases/Handlers/Boards/Commands/MoveBoardToAnotherProject/MoveBoardToAnotherProjectRequest.cs
@@ -5,8 +5,8 @@ namespace Vtodo.UseCases.Handlers.Boards.Commands.MoveBoardToAnotherProject
 {
     public class MoveBoardToAnotherProjectRequest : IRequest
     {
-        public int BoardId { get; set; }
+        public long BoardId { get; set; }
         
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
     }
 }
