+++ MainApp/Vtodo.UseCases/Handlers/ProjectsRoles/Commands/RevokeRole/RevokeRoleRequest.cs
@@ -5,7 +5,7 @@ namespace Vtodo.UseCases.Handlers.ProjectsRoles.Commands.RevokeRole
 {
     public class RevokeRoleRequest : IRequest
     {
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         
         public RevokeRoleDto RevokeRoleDto { get; set; } = null!;
     }
