+++ MainApp/Vtodo.UseCases/Handlers/Tasks/Commands/DeleteTask/DeleteTaskRequest.cs
@@ -1,10 +1,9 @@
 using MediatR;
-using Vtodo.UseCases.Handlers.Tasks.Dto;
 
 namespace Vtodo.UseCases.Handlers.Tasks.Commands.DeleteTask
 {
     public class DeleteTaskRequest : IRequest
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
     }
 }
