+++ MainApp/Vtodo.UseCases/Handlers/Tasks/Commands/MoveTaskToAnotherBoard/MoveTaskToAnotherBoardRequest.cs
@@ -4,7 +4,7 @@ namespace Vtodo.UseCases.Handlers.Tasks.Commands.MoveTaskToAnotherBoard
 {
     public class MoveTaskToAnotherBoardRequest : IRequest
     {
-        public int TaskId { get; set; }
-        public int NewBoardId { get; set; }
+        public long TaskId { get; set; }
+        public long NewBoardId { get; set; }
     }
 }
